Update price method for Beverage

+++ src/MilkTeaStore/IPriceStrategy.java
@@ -1,6 +1,6 @@
 package MilkTeaStore;
 
 public interface IPriceStrategy {
-	public int getPrice(double price);
+	public double getPrice(double price);
 	public int getPoint(double price);
 }
